extract generic lazyGCableObject utility function
