+++ src/platform/packages/shared/kbn-lazy-object/index.ts
@@ -10,6 +10,7 @@
 export { createLazyObjectFromFactories } from './src/create_lazy_object_from_factories';
 export { getLazyObjectMetrics } from './src/metrics';
 export { lazyObject } from './src/lazy_object';
+export { lazyGCableObject } from './src/lazy_gcable_object';
 export {
   createLazyObjectFromAnnotations,
   annotateLazy,
